requirements.txt
telethon
flask
flask-socketio
eventlet
python-dotenv
